requirements.txt
requests
randomcolor
matplotlib
python-dateutil
argparse